+++ src/core/base/wnComponent.js
@@ -188,12 +188,14 @@ module.exports = {
 		getFile: function (filePath,binary,cb)
 		{
 			var realPath = this.instanceOf('wnModule')?this.modulePath+filePath:filePath,
-				cmd = !cb ? 'readFileSync' : 'readFile',
-				_cb = cb ? function (err,file) {
+				cmd = !cb ? 'readFileSync' : 'readFile';
+				if (!fs.existsSync(realPath))
+					return (cb&&cb(false) == true);
+				var _cb = cb ? function (err,file) {
 					cb&&cb(!err ? ((binary===true) ? file : file.toString()) : false);
 				} : null,
 				file = fs[cmd](realPath,_cb);
-			return (file ? (binary===true) ? file : file.toString() : false);
+			return (file ? ((binary===true) ? file : file.toString()) : false);
 		},
 
 		/**
@@ -204,8 +206,10 @@ module.exports = {
 		getFileStat: function (filePath,cb)
 		{
 			var realPath = this.instanceOf('wnModule')?this.modulePath+filePath:filePath,
-				cmd = !cb ? 'statSync' : 'stat',
-				_cb = cb ? function (err,stat) {
+				cmd = !cb ? 'statSync' : 'stat';
+			if (!fs.existsSync(realPath))
+				return (cb&&cb(false) == true);
+			var _cb = cb ? function (err,stat) {
 					cb&&cb(!err ? stat : false);
 				} : null,
 				stat = fs[cmd](realPath,_cb);
@@ -307,22 +311,24 @@ module.exports = {
 				return _events[eventName];
 			else
 			{
+				if (!_eventsConfig[eventName] || !this.c)
+					return false;
 				var config = _eventsConfig[eventName] || {},
-					className = config.class;
-					config.id = eventName;
-					config.autoInit = false;
-					var evt = this.createClass(className,config);
-					evt.setParent(this);
-					evt.init();
-					if (hidden != false)
-					{
-						_events[eventName] = evt;
-						this.e[name]=function () { evt.push.apply(evt,arguments); };
-					} else
-					{
-						Object.defineProperty(_events[eventName],{ value: evt, enumerable: false });
-					}
-					return evt;
+					_class = config.class;
+				config.id = eventName;
+				config.autoInit = false;
+				var evt = this.createClass(_class,config);
+				evt.setParent(this);
+				evt.init();
+				if (hidden != false)
+				{
+					_events[eventName] = evt;
+					this.e[name]=function () { evt.push.apply(evt,arguments); };
+				} else
+				{
+					Object.defineProperty(_events[eventName],{ value: evt, enumerable: false });
+				}
+				return evt;
 			}
 		},
 
@@ -332,6 +338,14 @@ module.exports = {
 		getEvents: function ()
 		{
 			return _events;
+		},
+
+		/**
+		 * Get all defined configuration of events.
+		 */
+		getEventsConfig: function ()
+		{
+			return _eventsConfig;
 		},
 
 		/**