+++ src/core/base/wnConsole.js
@@ -102,22 +102,27 @@ module.exports = {
 		buildServer: function (serverPath)
 		{
 			var _sourcePath = path.resolve(cwd+sourcePath),
-				_serverPath = path.relative(cwd,serverPath),
+				_serverPath = serverPath,
 				relativeSourcePath = path.relative(serverPath,_sourcePath)+'/';
+			console.log("[*] Building new wnServer on `"+serverPath+"`");
 
 			if (!fs.existsSync(cwd+_serverPath))
 				return false;
 
+			console.log("[*] - Creating new `config.json` file.");
 			fs.writeFileSync(cwd+_serverPath+'/config.json',
 				fs.readFileSync(_sourcePath+'/default-config.json')
 			);
 
+			console.log("[*] - Creating new `index.js` file.");
 			var defaultIndex = fs.readFileSync(_sourcePath+'/default-index.js');
 			defaultIndex = new this.c.wnTemplate(defaultIndex).match({
 				sourcePath: relativeSourcePath.replace(/\\/g,'/'),
 				serverPath: _serverPath.replace(/\\/g,'/')
 			});
 			fs.writeFileSync(cwd+_serverPath+'/index.js',defaultIndex);
+
+			console.log('[*] New wnServer created.');
 
 			return true;
 		},
@@ -218,7 +223,7 @@ module.exports = {
 
 			this.e.log('Building wnServer from `'+serverPath+'`');
 
-			if (!fs.existsSync(this.modulePath+serverPath))
+			if (!fs.existsSync(this.modulePath+serverPath) || !fs.existsSync(this.modulePath+serverPath+'/config.json'))
 			{
 				this.e.log('Failed to load wnServer from path.');
 				return false;