+++ src/core/web/wnHttpRequest.js
@@ -88,34 +88,13 @@ module.exports = {
 		/**
 		 * Initializer
 		 */	
-		init: function ()
+		init: function (req,resp)
 		{
 			this.initialTime = +(new Date);
-			this.getEvent('end').setConfig({'source': this});
 
 			this.header = this.getConfig('header') || this.header;
-			this.info = this.getConfig('request');
-			this.response = this.getConfig('response');
-
-			_data = new Buffer(0);
-
-			if (this.response)
-			{
-				this.response.__write = this.response.write;
-				this.response.write = function (chunk, encoding)
-				{
-					if (chunk)
-						_data = Buffer.concat([_data,new Buffer(chunk,encoding)]);
-					self.response.__write.call(self.response,chunk,encoding);
-				};
-				this.response.__end = this.response.end;
-				this.response.end = function(chunk, encoding) {
-					if (chunk)
-						_data = Buffer.concat([_data,new Buffer(chunk,encoding)]);
-					self.response.__end.call(self.response,chunk,encoding);
-					self.e.end(self);
-			    };
-			}
+			this.info = req;
+			this.response = resp;
 
 			this.app = this.getParent();
 
@@ -137,11 +116,32 @@ module.exports = {
 			this.template = this.route ? this.route.template : false;
 
 			this.info.once('close',function () { self.e.end(self); });
+			this.info.once('end',function () { self.e.end(self); });
 			this.info.connection.setTimeout(this.lifeTime,function () {
 				self.info.connection.end();
 				self.e.end(self);
 				self.info.connection.destroy();
 			});
+
+			// _data = new Buffer(0);
+			// if (this.response)
+			// {
+			// 	this.response.__write = this.response.write;
+			// 	this.response.write = function (chunk, encoding)
+			// 	{
+			// 		if (chunk)
+			// 			_data = Buffer.concat([_data,new Buffer(chunk,encoding)]);
+			// 		self.response.__write.call(self.response,chunk,encoding);
+			// 	};
+			// 	this.response.__end = this.response.end;
+			// 	this.response.end = function(chunk, encoding) {
+			// 		if (chunk)
+			// 			_data = Buffer.concat([_data,new Buffer(chunk,encoding)]);
+			// 		self.response.__end.call(self.response,chunk,encoding);
+			// 		self.e.end(self);
+			//     };
+			// }
+
 			this.addListener('error',function (e,code,msg,fatal) {
 				this.err=true;
 				self.errorHandler(code,msg,fatal);
@@ -155,10 +155,6 @@ module.exports = {
 		 */	
 		run: function ()
 		{
-			this.once('run', function (e) {
-				if (self.cacheFilter())
-					e.stopPropagation=true;
-			});
 			this.once('run', function () {
 				if (self.template == '<file>')
 					self.publicHandler();
@@ -328,8 +324,12 @@ module.exports = {
 					res.setHeader(h,self.header[h]);
 				res.statusCode = self.code;
 				self.once('end',function () {
-					self.e.destroy(self);
-				})
+					self.app.once('closedRequest', function () {
+						self.e.destroy(self);
+					});
+					self.app.e.closedRequest(self);
+				});
+
 				res.end(self.data);
 			});
 		}