+++ src/wnBuild.js
@@ -380,3 +380,4 @@ wnBuild.prototype.makeDoc = function (className, docSource)
 		enumerable: false
 	});
 };
+